Misc
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1078 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/filed/backup.c
@@ -367,7 +367,7 @@ static int save_file(FF_PKT *ff_pkt, void *vjcr)
       /* 
        * Read the file data
        */
-      while ((sd->msglen=bread(&ff_pkt->bfd, rbuf, rsize)) > 0) {
+      while ((sd->msglen=(uint32_t)bread(&ff_pkt->bfd, rbuf, rsize)) > 0) {
 	 int sparseBlock = 0;
 
 	 /* Check for sparse blocks */
